Organize las carpetas del controller y las clases de persistencia
