+++ src/main/java/com/user_microservice/user/infrastructure/configuration/BeanConfiguration.java
@@ -9,12 +9,12 @@ import com.user_microservice.user.domain.spi.IUserModelPersistencePort;
 import com.user_microservice.user.domain.usecase.AuthenticationUseCase;
 import com.user_microservice.user.domain.usecase.RoleModelUseCase;
 import com.user_microservice.user.domain.usecase.UserModelUseCase;
-import com.user_microservice.user.infrastructure.output.jpa.adapter.RoleJpaAdapter;
-import com.user_microservice.user.infrastructure.output.jpa.adapter.UserJpaAdapter;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IRoleEntityMapper;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IUserEntityMapper;
-import com.user_microservice.user.infrastructure.output.jpa.repository.IRoleRepository;
-import com.user_microservice.user.infrastructure.output.jpa.repository.IUserRepository;
+import com.user_microservice.user.infrastructure.persistence.jpa.adapter.RoleJpaAdapter;
+import com.user_microservice.user.infrastructure.persistence.jpa.adapter.UserJpaAdapter;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IRoleEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IUserEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.repository.IRoleRepository;
+import com.user_microservice.user.infrastructure.persistence.jpa.repository.IUserRepository;
 import com.user_microservice.user.infrastructure.security.adapter.AuthenticationSecurityAdapter;
 import com.user_microservice.user.infrastructure.security.service.JwtService;
 import lombok.RequiredArgsConstructor;
