+++ src/main/java/com/user_microservice/user/infrastructure/http/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.user_microservice.user.infrastructure.input.controller;
+package com.user_microservice.user.infrastructure.http.controller;
 
 import com.user_microservice.user.application.dto.user_dto.UserRequest;
 import com.user_microservice.user.application.dto.user_dto.UserResponse;
