+++ src/main/java/com/user_microservice/user/infrastructure/persistence/jpa/adapter/RoleJpaAdapter.java
@@ -1,13 +1,13 @@
-package com.user_microservice.user.infrastructure.output.jpa.adapter;
+package com.user_microservice.user.infrastructure.persistence.jpa.adapter;
 
 import com.user_microservice.user.domain.exception.RoleNameNotFoundException;
 import com.user_microservice.user.domain.model.RoleModel;
 import com.user_microservice.user.domain.util.RoleName;
 import com.user_microservice.user.domain.spi.IRoleModelPersistencePort;
 import com.user_microservice.user.domain.util.Util;
-import com.user_microservice.user.infrastructure.output.jpa.entity.RoleEntity;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IRoleEntityMapper;
-import com.user_microservice.user.infrastructure.output.jpa.repository.IRoleRepository;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.RoleEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IRoleEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.repository.IRoleRepository;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
