+++ src/main/java/com/user_microservice/user/infrastructure/persistence/jpa/repository/IRoleRepository.java
@@ -1,7 +1,7 @@
-package com.user_microservice.user.infrastructure.output.jpa.repository;
+package com.user_microservice.user.infrastructure.persistence.jpa.repository;
 
 import com.user_microservice.user.domain.util.RoleName;
-import com.user_microservice.user.infrastructure.output.jpa.entity.RoleEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
