+++ src/main/java/com/user_microservice/user/infrastructure/persistence/jpa/repository/IUserRepository.java
@@ -1,6 +1,6 @@
-package com.user_microservice.user.infrastructure.output.jpa.repository;
+package com.user_microservice.user.infrastructure.persistence.jpa.repository;
 
-import com.user_microservice.user.infrastructure.output.jpa.entity.UserEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
