+++ src/main/java/com/user_microservice/user/infrastructure/security/adapter/AuthenticationSecurityAdapter.java
@@ -6,8 +6,8 @@ import com.user_microservice.user.domain.model.UserModel;
 import com.user_microservice.user.domain.security.IAuthenticationSecurityPort;
 import com.user_microservice.user.domain.spi.IRoleModelPersistencePort;
 import com.user_microservice.user.domain.util.Util;
-import com.user_microservice.user.infrastructure.output.jpa.entity.UserEntity;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IUserEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.UserEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IUserEntityMapper;
 import com.user_microservice.user.infrastructure.security.service.JwtService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
