+++ src/test/java/com/user_microservice/user/infrastructure/persistence/jpa/adapter/RoleJpaAdapterTest.java
@@ -1,11 +1,11 @@
-package com.user_microservice.user.infrastructure.output.jpa.adapter;
+package com.user_microservice.user.infrastructure.persistence.jpa.adapter;
 
 import com.user_microservice.user.domain.exception.RoleNameNotFoundException;
 import com.user_microservice.user.domain.model.RoleModel;
 import com.user_microservice.user.domain.util.RoleName;
-import com.user_microservice.user.infrastructure.output.jpa.entity.RoleEntity;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IRoleEntityMapper;
-import com.user_microservice.user.infrastructure.output.jpa.repository.IRoleRepository;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.RoleEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IRoleEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.repository.IRoleRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
