+++ src/test/java/com/user_microservice/user/infrastructure/persistence/jpa/adapter/UserJpaAdapterTest.java
@@ -1,9 +1,9 @@
-package com.user_microservice.user.infrastructure.output.jpa.adapter;
+package com.user_microservice.user.infrastructure.persistence.jpa.adapter;
 
 import com.user_microservice.user.domain.model.UserModel;
-import com.user_microservice.user.infrastructure.output.jpa.entity.UserEntity;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IUserEntityMapper;
-import com.user_microservice.user.infrastructure.output.jpa.repository.IUserRepository;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.UserEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IUserEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.repository.IUserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
