+++ src/test/java/com/user_microservice/user/infrastructure/security/adapter/AuthenticationSecurityAdapterTest.java
@@ -3,11 +3,9 @@ package com.user_microservice.user.infrastructure.security.adapter;
 import com.user_microservice.user.domain.exception.InvalidCredentialsException;
 import com.user_microservice.user.domain.model.RoleModel;
 import com.user_microservice.user.domain.model.UserModel;
-import com.user_microservice.user.domain.spi.IRoleModelPersistencePort;
 import com.user_microservice.user.domain.util.RoleName;
-import com.user_microservice.user.domain.util.Util;
-import com.user_microservice.user.infrastructure.output.jpa.entity.UserEntity;
-import com.user_microservice.user.infrastructure.output.jpa.mapper.IUserEntityMapper;
+import com.user_microservice.user.infrastructure.persistence.jpa.entity.UserEntity;
+import com.user_microservice.user.infrastructure.persistence.jpa.mapper.IUserEntityMapper;
 import com.user_microservice.user.infrastructure.security.service.JwtService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
